tidy up rest of entities helper setup; moved getMailsData to be with mailslurper
