+++ server-api/src/functional-api/notifications/community/invitations.it-spec.ts
@@ -1,8 +1,5 @@
-import { deleteMailSlurperMails } from '@utils/mailslurper.rest.requests';
-import {
-  deleteSpace,
-  updateSpaceSettings,
-} from '@functional-api/journey/space/space.request.params';
+import { deleteMailSlurperMails, getMailsData } from '@utils/mailslurper.rest.requests';
+import { updateSpaceSettings } from '@functional-api/journey/space/space.request.params';
 import { delay } from '@alkemio/tests-lib';
 import { users } from '@utils/queries/users-data';
 import {
@@ -10,8 +7,6 @@ import {
   inviteContributors,
 } from '@functional-api/roleset/invitations/invitation.request.params';
 import { TestUser } from '@alkemio/tests-lib';
-import { deleteOrganization } from '@functional-api/contributor-management/organization/organization.request.params';
-import { getMailsData } from '@src/types/entities-helper';
 import { PreferenceType } from '@generated/graphql';
 import { changePreferenceUser } from '@functional-api/contributor-management/user/user-preferences-mutation';
 import { TestScenarioFactory } from '@src/scenario/TestScenarioFactory';
@@ -56,8 +51,7 @@ const scenarioConfig: TestScenarioConfig = {
 beforeAll(async () => {
   await deleteMailSlurperMails();
 
-  baseScenario =
-    await TestScenarioFactory.createBaseScenario(scenarioConfig);
+  baseScenario = await TestScenarioFactory.createBaseScenario(scenarioConfig);
 
   await updateSpaceSettings(baseScenario.space.id, {
     membership: {
