+++ server-api/src/functional-api/notifications/community/notifications-registration.it-spec.ts
@@ -2,14 +2,14 @@ import {
   createUser,
   deleteUser,
 } from '@functional-api/contributor-management/user/user.request.params';
-import { UniqueIDGenerator } from '@alkemio/tests-lib';;
-const uniqueId = UniqueIDGenerator.getID();
-import { deleteMailSlurperMails } from '@utils/mailslurper.rest.requests';
+import { UniqueIDGenerator } from '@alkemio/tests-lib';
+import { deleteMailSlurperMails, getMailsData } from '@utils/mailslurper.rest.requests';
 import { users } from '@utils/queries/users-data';
-import { getMailsData } from '@src/types/entities-helper';
 import { changePreferenceUser } from '@functional-api/contributor-management/user/user-preferences-mutation';
 import { PreferenceType } from '@generated/graphql';
 import { delay } from '@alkemio/tests-lib';
+
+const uniqueId = UniqueIDGenerator.getID();
 
 let userName = '';
 let userId = '';
