+++ server-api/src/functional-api/notifications/messaging/mention-user.it-spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { deleteMailSlurperMails } from '@utils/mailslurper.rest.requests';
+import { deleteMailSlurperMails, getMailsData } from '@utils/mailslurper.rest.requests';
 import { delay } from '@alkemio/tests-lib';
 import { TestUser } from '@alkemio/tests-lib';
 import { UniqueIDGenerator } from '@alkemio/tests-lib';
@@ -8,7 +8,6 @@ import { createPostOnCallout } from '@functional-api/callout/post/post.request.p
 import { PreferenceType } from '@generated/alkemio-schema';
 import { changePreferenceUser } from '@functional-api/contributor-management/user/user-preferences-mutation';
 import { sendMessageToRoom } from '@functional-api/communications/communication.params';
-import { getMailsData } from '@src/types/entities-helper';
 import { TestScenarioFactory } from '@src/scenario/TestScenarioFactory';
 import { OrganizationWithSpaceModel } from '@src/scenario/models/OrganizationWithSpaceModel';
 import { TestScenarioConfig } from '@src/scenario/config/test-scenario-config';
