+++ server-api/src/utils/kratos/verify-in-kratos-or-fail.ts
@@ -1,8 +1,8 @@
 import request from 'supertest';
-import { delay } from '../../../../lib/src/utils/delay';
 import { getMails } from '../mailslurper.rest.requests';
 import { Configuration, IdentityApi, FrontendApi } from '@ory/kratos-client';
 import { testConfiguration } from '@src/config/test.configuration';
+import { delay } from '@alkemio/tests-lib';
 
 /***
  * Verification flow on v0.8.0-alpha3
